refactor: simple content
